requirements.txt
praw~=7.6.0
requests~=2.28.1
Pillow~=9.2.0
python-dotenv~=0.21.0